requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
openai>=1.54.0
requests>=2.32.0
duckduckgo-search>=8.0.0
pyyaml>=6.0.2

